separated the routes and the code that actually runs on the route
